# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: WebCon API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspace:
    get:
      operationId: listWorkspaceItems
      tags: [workspace]
      summary: List workspace items
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Workspace items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceItem"
    post:
      operationId: createWorkspaceItem
      tags: [workspace]
      summary: Create a workspace item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceItemBody"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceItem"
  /workspace/{id}:
    patch:
      operationId: updateWorkspaceItem
      tags: [workspace]
      summary: Update a workspace item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceItem"
    delete:
      operationId: deleteWorkspaceItem
      tags: [workspace]
      summary: Delete a workspace item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /schedule:
    get:
      operationId: listScheduleSessions
      tags: [schedule]
      summary: List study schedule sessions
      responses:
        "200":
          description: Schedule sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSession"
    post:
      operationId: createScheduleSession
      tags: [schedule]
      summary: Create a schedule session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSession"
  /schedule/{id}:
    patch:
      operationId: updateScheduleSession
      tags: [schedule]
      summary: Update a schedule session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleSession"
    delete:
      operationId: deleteScheduleSession
      tags: [schedule]
      summary: Delete a schedule session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/tasks:
    post:
      operationId: createProjectTask
      tags: [projects]
      summary: Add a task to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTask"
  /projects/{id}/tasks/{taskId}:
    patch:
      operationId: updateProjectTask
      tags: [projects]
      summary: Update a project task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectTask"
    delete:
      operationId: deleteProjectTask
      tags: [projects]
      summary: Delete a project task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    WorkspaceItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        agentId:
          type: ["integer", "null"]
        conversationId:
          type: ["integer", "null"]
        type:
          type: string
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
        starred:
          type: boolean
        subject:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - agentId
        - conversationId
        - type
        - title
        - content
        - pinned
        - starred
        - subject
        - createdAt
        - updatedAt
    CreateWorkspaceItemBody:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        content:
          type: string
        agentId:
          type: integer
        conversationId:
          type: integer
        subject:
          type: string
        pinned:
          type: boolean
        starred:
          type: boolean
      required:
        - type
        - title
    UpdateWorkspaceItemBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
        starred:
          type: boolean
        subject:
          type: string
    ScheduleSession:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        title:
          type: string
        subject:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        duration:
          type: integer
        type:
          type: string
        completed:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - agentId
        - agentName
        - title
        - subject
        - date
        - duration
        - type
        - completed
        - notes
        - createdAt
    CreateScheduleSessionBody:
      type: object
      properties:
        title:
          type: string
        agentId:
          type: integer
        subject:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        type:
          type: string
        notes:
          type: string
      required:
        - title
        - date
    UpdateScheduleSessionBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        type:
          type: string
        completed:
          type: boolean
        notes:
          type: string
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        title:
          type: string
        subject:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
          format: date-time
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ProjectTask"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - agentId
        - agentName
        - title
        - subject
        - type
        - status
        - dueDate
        - tasks
        - createdAt
        - updatedAt
    ProjectTask:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        completed:
          type: boolean
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - completed
        - dueDate
        - createdAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        agentId:
          type: integer
        subject:
          type: string
        type:
          type: string
        dueDate:
          type: string
          format: date-time
      required:
        - title
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        type:
          type: string
        status:
          type: string
        dueDate:
          type: string
          format: date-time
    CreateProjectTaskBody:
      type: object
      properties:
        title:
          type: string
        dueDate:
          type: string
          format: date-time
      required:
        - title
    UpdateProjectTaskBody:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean
        dueDate:
          type: string
          format: date-time
...